add data persistence by Room

+++ app/src/main/java/com/example/mitweather/data/model/Clouds.kt
@@ -1,5 +1,5 @@
 package com.example.mitweather.data.model
 
-data class Clouds(
-    val all: Int
+data class Clouds (
+    val all: Long
 )
